+++ api/tests_bdd/steps/schemas.py
@@ -1,15 +1,15 @@
 import json
 
 from behave import given, then, when
-from utils.package_import import import_package
 
 from classes.dto import DTO
 from classes.schema import Factory
 from config import Config
+from core.repository.dmss.TemplateRepositoryFromDMSS import TemplateRepositoryFromDMSS
+from services.data_modelling_document_service import explorer_api
 from utils.data_structure.compare import pretty_eq
-from utils.helper_functions import schemas_location
 from utils.logging import logger
-from services.data_modelling_document_service import document_api, explorer_api, package_api, datasource_api
+from utils.package_import import import_package
 
 
 @given("data modelling tool templates are imported")
@@ -25,17 +25,6 @@ def step_impl_2(context, uid: str, data_source_id: str):
     document: DTO = DTO(uid=uid, data=json.loads(context.text))
     response = explorer_api.add_raw(data_source_id, document.to_dict())
     print(response)
-
-
-class TemplateRepositoryFromDMSS:
-    def get(self, template_type: str):
-        return self[template_type]
-
-    def __getitem__(self, template_type: str) -> dict:
-        data_source, *rest = template_type.split("/")
-        template_type = "/".join(rest)
-        document = document_api.get_by_path(data_source, template_type)
-        return document["document"]
 
 
 @when('I create a Python class from the template "{template_name}"')
